refactor: reorganize product filter logic into dedicated search params modules
