+++ lib/search-params/product-filters.server.ts
@@ -1,0 +1,4 @@
+import { createLoader } from "nuqs/server";
+import { productFiltersParams } from "./product-filters";
+
+export const loadProductFilters = createLoader(productFiltersParams);
