+++ lib/search-params/product-filters.ts
@@ -1,0 +1,11 @@
+import { sortValues } from "@/constants";
+import { parseAsArrayOf, parseAsString, parseAsStringLiteral } from "nuqs";
+
+export const productFiltersParams = {
+  sort: parseAsStringLiteral(sortValues).withDefault("curated"),
+  minPrice: parseAsString.withOptions({ clearOnDefault: true }).withDefault(""),
+  maxPrice: parseAsString.withOptions({ clearOnDefault: true }).withDefault(""),
+  tags: parseAsArrayOf(parseAsString)
+    .withOptions({ clearOnDefault: true })
+    .withDefault([] as string[]),
+};
